Fix CacheBase

+++ Classes/PHPExcel/CachedObjectStorage/CacheBase.php
@@ -152,7 +152,9 @@ abstract class PHPExcel_CachedObjectStorage_CacheBase {
      */
 	public function deleteCacheData($pCoord) {
 		if ($pCoord === $this->_currentObjectID) {
-			$this->_currentObject->detach();
+			if (isset($this->_currentObject)) {
+				$this->_currentObject->detach();
+			}
 			$this->_currentObjectID = $this->_currentObject = null;
 		}
